fastapi-server/pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.5.3",
    "fastapi>=0.116.1",
    "flower>=2.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-postgres>=0.0.15",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "uvicorn>=0.35.0",
    "deepagents>=0.0.5",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[google-genai]>=0.3.27",
    "langgraph-api>=0.2.3",
    "langgraph-sdk>=0.1.61",
]
